Tests/LibWeb/Text/expected/css/computed-font-style-with-relative-lengths-in-calc.txt

oblique calc(57.29578deg * sign(1em - 1px))
